refactoring action-creators (Edit Profile)
