+++ EventsExpress/ClientApp/src/components/Route guard/401.js
@@ -7,7 +7,6 @@ import { setRegisterError } from '../../actions/register'
 import { setEditUsernameError } from '../../actions/EditProfile/editUsername'
 import { setEditGenderError } from '../../actions/EditProfile/EditGender'
 import { setEditBirthdayError } from '../../actions/EditProfile/editBirthday'
-import { setAvatarError } from '../../actions/EditProfile/change-avatar'
 import { getUsersError } from '../../actions/users'
 import { getEventError } from '../../actions/event-item-view'
 import { setRolesError } from '../../actions/roles'
