+++ EventsExpress/ClientApp/src/containers/categories/SelectCategories.js
@@ -1,7 +1,7 @@
 ﻿import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import SelectCategories from '../../components/SelectCategories/SelectCategories';
-import setUserCategory from '../../actions/EditProfile/addUserCategory';
+import setUserCategory from '../../actions/EditProfile/userCategory-add-action';
 import get_categories from '../../actions/category/category-list-action';
 
 class SelectCategoriesWrapper extends Component {
