+++ EventsExpress/ClientApp/src/reducers/user.js
@@ -1,12 +1,12 @@
 import initialState from '../store/initialState';
 import { SET_USER } from '../actions/login-action';
 import { SET_LOGOUT } from '../actions/logout-action';
-import { addUserCategory } from '../actions/EditProfile/addUserCategory';
+import { addUserCategory } from '../actions/EditProfile/userCategory-add-action';
 import { addUserNotificationType } from '../actions/EditProfile/addUserNotificationType';
 import { editBirthday } from '../actions/EditProfile/editBirthday';
 import { editGender } from '../actions/EditProfile/EditGender';
 import { editUsername } from '../actions/EditProfile/editUsername';
-import { changeAvatar } from '../actions/EditProfile/change-avatar';
+import { changeAvatar } from '../actions/EditProfile/avatar-change-action';
 import { authenticate } from '../actions/authentication-action';
 
 export const reducer = (state = initialState.user, action) => {
