+++ EventsExpress/ClientApp/src/services/CategoryService.js
@@ -4,31 +4,16 @@ const baseService = new EventsExpressService();
 
 export default class CategoryService {
 
-    getAllCategories = () => baseService.getResource('category/all');
+    getAllCategories = () => baseService.getResourceNew('category/all');
 
-    setCategory = async (data) => {
-        const res = await baseService.setResource('category/create', {
+    setCategory = data => baseService.setResource('category/create', {
             name: data.name
         });
-        return !res.ok
-            ? { error: await res.text() }
-            : res;
-    }
 
-    editCategory = async (data) => {
-        const res = await baseService.setResource('category/edit', {
+    editCategory = data => baseService.setResource('category/edit', {
             id: data.id,
             name: data.name
         });
-        return !res.ok
-            ? { error: await res.text() }
-            : res;
-    }
 
-    setCategoryDelete = async (data) => {
-        const res = await baseService.setResource(`category/delete/${data}`);
-        return !res.ok
-            ? { error: await res.text() }
-            : res;
-    }
+    setCategoryDelete = data => baseService.setResource(`category/delete/${data}`);
 }
